==> Bengaluru/08C-Model-Spatial-Relation-Bengaluru-All-PM2.5.ipynb <==
import pickle
from glob import glob

import numpy as np
import pandas as pd

import datetime

def extract_date_from_file_name(FILE_NAME):
    return datetime.datetime.strptime(FILE_NAME.split('/')[-1].split('.')[1:][0][1:], "%Y%j")

import configparser


config = configparser.ConfigParser()
config.read("./config.ini")

PATH = config['path']['hdf_path_2020'] + "*"
FILE_LIST = glob(PATH)

pkl_obj = pickle.load(open("bengaluru_geodata.pkl", "rb"))
BENGALURU_B_MIN_LON, BENGALURU_B_MAX_LON = pkl_obj['min_lon'], pkl_obj['max_lon']
BENGALURU_B_MIN_LAT, BENGALURU_B_MAX_LAT = pkl_obj['min_lat'], pkl_obj['max_lat']
boundary = pkl_obj['boundary']
mask = pkl_obj['bengaluru_mask'] 

SDS_NAME = "Optical_Depth_047"

import matplotlib.pyplot as plt
from mpl_toolkits.basemap import Basemap, cm

def plot_bengaluru_pm25_map(name, scaled_data, date_time, save=False, should_mask=True):

    add_text = ""
    
    plt.figure(figsize=(20, 10))

    pad = 0.01

    m = Basemap(projection='cyl', resolution='l', llcrnrlat=BENGALURU_B_MIN_LAT-pad, urcrnrlat=BENGALURU_B_MAX_LAT+pad, llcrnrlon=BENGALURU_B_MIN_LON-pad, urcrnrlon=BENGALURU_B_MAX_LON+pad , suppress_ticks=True)

    m.readshapefile("../Municipal_Spatial_Data/Bengaluru/Bengaluru",'Bengaluru', linewidth=1)

    m.drawparallels(np.arange(-90., 120., 0.05), labels=[1, 0, 0, 0], fontsize=20)
    m.drawmeridians(np.arange(-180., 181., 0.05), labels=[0, 0, 0, 1],  fontsize=20)
    x, y = m(mcd19a2_longitude[mask], mcd19a2_latitude[mask])

    if should_mask:
        sc = m.scatter(x, y, c=scaled_data[mask], s=1650, cmap=plt.cm.jet,
                        edgecolors='white', linewidth=0, marker='s')
    else:
        add_text = "_pm25"
        sc = m.scatter(x, y, c=scaled_data, s=1650, cmap=plt.cm.jet,
                        edgecolors='white', linewidth=0, marker='s')

    cb = m.colorbar()
    cb.set_label('PM2.5', size=20)
    cb.ax.tick_params(labelsize=18)

    plotTitle = name[:-4]
    plt.title('{0}\n {1}\n Bengaluru, India\n {2}'.format(plotTitle, "PM2.5", date_time.strftime("%Y-%m-%d")), fontsize=28)
    fig = plt.gcf()

#     plt.show()
    
    if save:
        pngfile = 'images/{0}{1}.png'.format(plotTitle, add_text)
        fig.savefig(pngfile, bbox_inches="tight", dpi=400)

def hypotenuse_distance(x, y):
    return np.sqrt(x**2 + y**2)

def model_predict(X, regressor, scaler):    
    if scaler:
        try:
            sc_X, sc_y = scaler
            y_pred = sc_y.inverse_transform(regressor.predict(sc_X.transform(X)))
        except Exception as e:
            poly_reg = scaler
            y_pred = regressor.predict(poly_reg.transform(X))
    else:
        y_pred = regressor.predict(X)
    distances = np.array(list(map(lambda x: hypotenuse_distance(x[1], x[2]), X)))
    return np.array(list(zip(y_pred, distances)))

def get_weighted_value(y):
    weights = 1 / pow(y[:,1], 2)
    return sum(weights * y[:, 0]) / sum(weights)
    
def mix(y_list):
    return np.array(list(map(lambda x: get_weighted_value(x), y_list)))

def get_nearest_3x3_grid(data, x, y):
    
    if x < 1:
        x += 1
    if x > data.shape[0]-2:
        x -= 2
    if y < 1:
        y += 1
    if y > data.shape[1]-2:
        y -= 2  
    
    three_by_three = data[x-1:x+2,y-1:y+2]
    three_by_three = three_by_three.astype(float)
    
    not_nans = np.count_nonzero(~np.isnan(three_by_three))
    
    if not_nans == 0:
        return {
            "x": x,
            "y": y,
        }
    else:
        three_by_three_average = np.nanmean(three_by_three)
        three_by_three_std = np.nanstd(three_by_three)
        three_by_three_median = np.nanmedian(three_by_three)
        
        return {
            "x": x,
            "y": y,
            "data": three_by_three,
            "average": three_by_three_average,
            "std": three_by_three_std,
            "median": three_by_three_median
        }    

BENGALURU_GEO_OBJ = pickle.load(open("./bengaluru_geodata.pkl", "rb"))
BENGALURU_MASK = BENGALURU_GEO_OBJ['bengaluru_mask']
BENGALURU_COORDS = BENGALURU_GEO_OBJ['bengaluru_coords']

DECIMAL_PLACES = 7
PROCESS_ML_DATA = True

CITY = "Bengaluru"

india_stations_df = pd.read_pickle('../2015-2020-pm25/india_stations.pkl')
stations = india_stations_df[india_stations_df['City'] == CITY]['StationId'].values

mcd19a2_obj = pickle.load(open("./mcd19a2.pkl", "rb"))
mcd19a2_longitude, mcd19a2_latitude = mcd19a2_obj['longitude'], mcd19a2_obj['latitude']

model_dir = "spt_models"

import warnings
warnings.filterwarnings("ignore")

def aod2pm25_model_predict(X, regressor, scaler):    
    if scaler:
        try:
            sc_X, sc_y = scaler
            y_pred = sc_y.inverse_transform(regressor.predict(sc_X.transform(X)))
        except Exception as e:
            poly_reg = scaler
            y_pred = regressor.predict(poly_reg.transform(X))
    else:
        y_pred = regressor.predict(X)
    return y_pred

aod2pm25_model = pickle.load(open(config['convert']["aod_pm25"], "rb"))

from pyhdf import SD

def perform_task(idx):
    
    empty_flag = 0
    
    FILE_NAME = FILE_LIST[idx]
    NAME = FILE_NAME.split('/')[-1]
    date_time = extract_date_from_file_name(FILE_NAME)
    
    # ------------------ Data Loading Begin -------------------------

    hdf = SD.SD(FILE_NAME)
    sds = hdf.select(SDS_NAME)
    data = sds.get()

    attributes = sds.attributes()
    scale_factor = attributes['scale_factor']
    fv = attributes['_FillValue']

    data = data.astype(float)
    data[data == fv] = np.nan
    data = np.nanmean(data, axis=0)

    scaled_data = data * scale_factor
    
    # ------------------ Data Loading End -------------------------

#     plot_bengaluru_pm25_map(NAME, scaled_data, date_time, save=True);
    
    # ------------------ Interpolation Begin -------------------------

    y_sp_interpolate = []

    for i, station_id in enumerate(stations):
        best = pickle.load(open(f"{model_dir}/{station_id}_spt.pkl", "rb"))

        fix_station_lat, fix_station_lon = best['station_id']['nearest_lat'], best['station_id']['nearest_lon']
        x_coord = best['station_id']['x_coord']
        y_coord = best['station_id']['y_coord']

        regressor = best['regressor']
        scaler = best['scaler']

        try:
            fix_station_aod = get_nearest_3x3_grid(scaled_data, x_coord, y_coord)['average'].round(3)
            X_sp_interpolate_sub = np.array(list(zip([fix_station_aod for k in range(BENGALURU_COORDS.shape[0])],
                             (fix_station_lat - BENGALURU_COORDS[:,0]),
                             (fix_station_lon - BENGALURU_COORDS[:,1]),
                            )
                        ))

            y_sp_interpolate_sub = model_predict(X_sp_interpolate_sub, regressor, scaler)
            y_sp_interpolate.append(y_sp_interpolate_sub)
        except KeyError :
            pass
        
    try:
        y_sp_interpolate = np.array(y_sp_interpolate)
        y_sp_interpolate = y_sp_interpolate.transpose(1, 0, 2)
        y_sp_interpolate = mix(y_sp_interpolate).round(3)
        y_sp_interpolate = y_sp_interpolate.reshape(-1,1)
        
        y_pm25_predictions = aod2pm25_model_predict(y_sp_interpolate, aod2pm25_model['regressor'], aod2pm25_model['scaler']).reshape(-1,1)
        
        plot_bengaluru_pm25_map(NAME, y_pm25_predictions, date_time, save=True, should_mask=False);
    except Exception as e:
        pass

import multiprocessing as mp

pool = mp.Pool(mp.cpu_count())

pool.map(perform_task, [idx for idx in range(len(FILE_LIST))])
pool.close()
